big update

+++ src/utils/types.rs
@@ -1,30 +1,30 @@
 use std::ops::Range;
 
 pub enum Types {
-    One(u32),
-    Range(Range<u32>),
+    One(usize),
+    Range(Range<usize>),
 }
 impl Types {
-    pub fn to_vec(&self) -> Vec<u32> {
+    pub fn to_vec(&self) -> Vec<usize> {
         match self {
             Types::One(i) => vec![*i],
             Types::Range(r) => r.to_owned().collect(),
         }
     }
-    pub fn to_range(&self) -> Range<u32> {
+    pub fn to_range(&self) -> Range<usize> {
         match self {
             Types::One(i) => *i..*i + 1,
             Types::Range(r) => r.clone(),
         }
     }
 }
-impl From<Range<u32>> for Types {
-    fn from(value: Range<u32>) -> Self {
+impl From<Range<usize>> for Types {
+    fn from(value: Range<usize>) -> Self {
         Types::Range(value)
     }
 }
-impl From<u32> for Types {
-    fn from(value: u32) -> Self {
+impl From<usize> for Types {
+    fn from(value: usize) -> Self {
         Self::One(value)
     }
 }
